requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.4b2
bleach==6.0.0
boto3==1.28.40
botocore==1.31.40
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
diffusers==0.10.0
docker-pycreds==0.4.0
einops==0.3.0
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.16.4
hydra-core==1.1.1
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.13.1
kiwisolver==1.4.4
kornia==0.6.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
lit==16.0.6
lvis==0.5.3 # @ https://github.com/lvis-dataset/lvis-api/archive/master.zip#sha256=fc9bdf207e41a883e046f4343759d2966c65e5d946fc39a88fdf689ccd1c78c8
Markdown==3.4.4
MarkupSafe==2.1.3
# mask2former -e third_party/Mask2Former/
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
# -e git+https://github.com/NVlabs/ODISE@03d19069c563567240e67fe9db707acf6f573260#egg=odise
# odise -e third_party/odise/
omegaconf==2.1.1
open-clip-torch==2.0.2
opencv-contrib-python==4.8.0.76
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
python-panopticon
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
# pydensecrf -e git+https://github.com/lucasb-eyer/pydensecrf.git
pyDeprecate==0.3.1
pydot==1.4.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.4.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stable-diffusion-sdkit==2.1.3
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
terminado==0.17.1
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.8.30
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
# torch==2.0.1+cu118
# torchaudio==2.0.2+cu118
torchmetrics==0.6.0
# torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.26.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wandb==0.15.9
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.7
wget==3.2
widgetsnbextension==4.0.8
wrapt==1.15.0
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2